+++ source/helper.js
@@ -37,7 +37,9 @@ Helper.queue = queue((task, callback) => {
   Jwt.getJWT(task.opt)
     .then((jwtToken) => {
       const nonce = Date.now();
-      const message = `${nonce}${config.getApiUrl(task.opt.useSandbox, task.opt.url)}${task.path}${JSON.stringify(task.rawBody)}`;
+      const message = `${nonce}${config.getApiUrl(task.opt.useSandbox, task.opt.url)}${
+        task.path
+      }${JSON.stringify(task.rawBody)}`;
       const hmac = Hmac.generate(task.opt.privateKey, message);
 
       // make the request to blockchainiz